SSZ (#878)
* fix: local `prometheus.yaml` and allow zero `GenesisEpoch`

* no slashing 😱

* fix: removed incorrect `ValidatorIndex == 0` check

* print consensus time

* remove git --fetch from Dockerfile

* dont fetch sync duties

* upgrade ssv-spec dependency

* added `patches` directory to .gitignore

* move `SignedPartialSignatureMessage` from package `ssv` to `types`

* copy/paste part 1

* copy/paste part 2

* gopatch

* copy/paste part 3

* part 4

* delete deprecated beacon_mock.go

* SSV message creation in tests

* more refactors

* adapt to new NewMsgID with DomainType

* more fixes

* fix: all unit tests

* fix: unit test

* fix: all unit tests

* updated tests.json

* fix: most spec tests

* fix: all spec tests ✅

* removed useless log

* fix: lint

* fix: (ssv/validator) log pubkey

* deploy to v3-1..4

* fix: NETWORK_ID should be hex

* TEMPORARY LOGGING TO REVERT LATER

* TEMPORARY LOGGING TO REVERT LATER

* TEMPORARY LOG TO REVERT LATER

* fix: missing Slot in PartialSignatureMessages construction

* fix: don't fetch sync committee contributions for non-aggregators

* TEMPORARY LOGGING TO REVERT LATER

* update ssv-spec to fork with SSZ32Bytes fix

* swap databases (after highest decided decoding errors)

* deploy to 9..12

* TEMPORARY ROUTE

* TEMPORARY ROUTE

* Revert "TEMPORARY ROUTE"

This reverts commit b5fbb2776953935fed29e218a0dad881931c86e9.

* Revert "TEMPORARY ROUTE"

This reverts commit cee426b2d2aa06e1ac9843a329907e789bd3d4c5.

* Revert "TEMPORARY LOGGING TO REVERT LATER"

This reverts commit e71760f100f33783666b2f807a098718470a4208.

* Revert "TEMPORARY LOG TO REVERT LATER"

This reverts commit 5f678692b22667e40bc5ee9b9cd51bc0c28dcea1.

* Revert "TEMPORARY LOGGING TO REVERT LATER"

This reverts commit 0c1039533f0002a00d67bbaf224a0bde9584440e.

* Revert "TEMPORARY LOGGING TO REVERT LATER"

This reverts commit d066ea86b4dec181f083187e5c442fd805cba765.

* revert logging fixes

* revert prod deployment changes

* added format_test.go

* fix: deployments

* fix: NETWORK_ID for stage

+++ exporter/metrics.go
@@ -29,6 +29,6 @@ func init() {
 // ReportOperatorIndex reporting of new or exist operators
 func ReportOperatorIndex(logger *zap.Logger, op *registrystorage.OperatorData) {
 	pkHash := fmt.Sprintf("%x", sha256.Sum256(op.PublicKey))
-	metricOperatorIndex.WithLabelValues(pkHash, strconv.FormatUint(uint64(op.ID), 10)).Set(float64(op.ID))
-	logger.Debug("report operator", zap.String("pkHash", pkHash), zap.Uint64("id", uint64(op.ID)))
+	metricOperatorIndex.WithLabelValues(pkHash, strconv.FormatUint(op.ID, 10)).Set(float64(op.ID))
+	logger.Debug("report operator", zap.String("pkHash", pkHash), zap.Uint64("id", op.ID))
 }
