+++ logging/fields/fields.go
@@ -152,7 +152,7 @@ func SyncResults(msgs protocolp2p.SyncResults) zapcore.Field {
 }
 
 func OperatorID(operatorId spectypes.OperatorID) zap.Field {
-	return zap.Uint64("operator-id", uint64(operatorId))
+	return zap.Uint64("operator-id", operatorId)
 }
 
 func OperatorIDStr(operatorId string) zap.Field { //todo cleanup it
